Sucessfully serialize and de-serialize connections
